improve/pac/requirements.txt
flamingo_pytorch
ftfy 
regex 
transformers==4.5.1
git+https://github.com/openai/CLIP.git
accelerate
